requirements.txt
requests==2.31.0
pandas==2.1.0
streamlit==1.26.0
plotly==5.16.1
scipy==1.11.2
numpy==1.24.3
python-dotenv==1.0.0
SQLAlchemy==2.0.20
aiohttp==3.8.5